gofmt アドレスは64bitで取り扱うようにする

+++ os/tool/asm/assembler/instruction/mnemonic.go
@@ -6,15 +6,15 @@ import "io"
 type Mnemonic interface {
 
 	// 自分自身の命令サイズを返す
-	Size() int
+	Size() int64
 
 	// ラベルの解決
 	//
 	// @param table --- ラベルテーブル
 	//
 	// @return エラー ラベルが見つからない場合
-	Relocate(table map[string]int) error
+	Relocate(table map[string]int64) error
 
 	// オペレーションをバイナリとして出力
-	Write(w io.Writer) (int, error)
+	Write(w io.Writer) (int64, error)
 }
